refactor(commands): extract bug/update tests to sibling file
